Move all logic to acl file to keep it in one place
